# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Gestion parc auto ",
        "description": "API pour la gestion d'un parc auto",
        "contact": {
            "email": "bistephaneta@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/achats": {
            "get": {
                "tags": [
                    "Achats"
                ],
                "summary": "Affiche tous les achats",
                "description": "Retourne tous les achats",
                "responses": {
                    "200": {
                        "description": "opération réussi"
                    }
                }
            },
            "post": {
                "tags": [
                    "Achats"
                ],
                "summary": "ajoute un achat",
                "description": "Retourne toutes les information d'un achat",
                "parameters": [
                    {
                        "name": "nom",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "prenoms",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contact",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/v1/clients": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Affiche tous les clients",
                "description": "Retourne tous les clients",
                "responses": {
                    "200": {
                        "description": "opération réussi"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "ajoute un client",
                "description": "Retourne toutes les information du client",
                "parameters": [
                    {
                        "name": "nom",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "prenoms",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contact",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/v1/clients/{id}": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "affiche un client",
                "description": "Retourne les information d'un client",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Mise à jour d'un client ",
                "description": "Retourne la mise à jour des informations d'un client",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "nom",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "prenoms",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "contact",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "supprime un client existant",
                "description": "Supprime un client ",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/v1/vehicules": {
            "get": {
                "tags": [
                    "Vehicules"
                ],
                "summary": "Affiche tous les vehicules",
                "description": "Retourne tous les clients",
                "responses": {
                    "200": {
                        "description": "opération réussi"
                    }
                }
            },
            "post": {
                "tags": [
                    "Vehicules"
                ],
                "summary": "Ajoute un vehicule",
                "description": "Retourne les informations d'un vehicule",
                "parameters": [
                    {
                        "name": "marque",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "modele",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "annee",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/v1/vehicules/{id}": {
            "get": {
                "tags": [
                    "Vehicules"
                ],
                "summary": "Affiche un vehicule",
                "description": "Retourne les informations d'un vehicule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vehicules"
                ],
                "summary": "Mise à jour d'un vehicule ",
                "description": "Retourne la mise à jour des informations d'un vehicule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "202": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vehicules"
                ],
                "summary": "Supprime un vehicule",
                "description": "Supprime un vehicule",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Project id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    }
}